requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.5
pip==18.1
pkg-resources==0.0.0
prometheus-client==0.7.1
python-gitlab==2.2.0
requests==2.23.0
setuptools==40.8.0
urllib3==1.25.8
wheel==0.34.2
yarl==1.4.2
